move payment  history modal to its own file
